extract duplicated code between shell exercises
